User authentication + bug fixes

+++ src/controllers/foldersController.ts
@@ -10,8 +10,11 @@ export const getFolders: RequestHandler = async (
   req: Request,
   res: Response
 ): Promise<void> => {
-  const folders = await Folder.find({}).sort({ createdAt: -1 });
-  res.status(200).json(folders);
+  const { user } = req;
+  if (user) {
+    const folders = await Folder.find({ user }).sort({ createdAt: -1 });
+    res.status(200).json(folders);
+  }
 };
 
 // get a folder
@@ -20,20 +23,24 @@ export const getFolder: RequestHandler = async (
   req: Request,
   res: Response
 ): Promise<void> => {
-  const { id } = req.params;
-  if (!mongoose.Types.ObjectId.isValid(id)) {
-    res.status(400).json({ error: "No such Folder" });
-    return;
+  const { user } = req;
+
+  if (user) {
+    const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      res.status(400).json({ error: "No such Folder" });
+      return;
+    }
+
+    const folder = await Folder.findOne({ _id: id, user });
+
+    if (!folder) {
+      res.status(404).json({ error: "No such Folder" });
+      return;
+    }
+
+    res.status(200).json(folder);
   }
-
-  const folder = await Folder.findById(id);
-
-  if (!folder) {
-    res.status(404).json({ error: "No such Folder" });
-    return;
-  }
-
-  res.status(200).json(folder);
 };
 
 // create a folder
@@ -42,16 +49,19 @@ export const postFolder: RequestHandler = async (
   req: Request,
   res: Response
 ): Promise<void> => {
-  if ("name" in req.body && "notes" in req.body) {
-    const { name, notes } = req.body as FolderObj;
-    try {
-      const folder = await Folder.create({ name, notes });
-      res.status(201).json(folder);
-    } catch (err) {
-      if (err instanceof Error) res.status(400).json({ error: err.message });
+  const { user } = req;
+  if (user) {
+    if ("name" in req.body && "notes" in req.body) {
+      const { name, notes } = req.body as FolderObj;
+      try {
+        const folder = await Folder.create({ name, notes, user });
+        res.status(201).json(folder);
+      } catch (err) {
+        if (err instanceof Error) res.status(400).json({ error: err.message });
+      }
+    } else {
+      res.json({ error: "Folder name and/or notes not found" });
     }
-  } else {
-    res.json({ error: "Folder name and/or notes not found" });
   }
 };
 
@@ -61,22 +71,25 @@ export const deleteFolder: RequestHandler = async (
   req: Request,
   res: Response
 ): Promise<void> => {
-  const { id } = req.params;
+  const { user } = req;
+  if (user) {
+    const { id } = req.params;
 
-  if (!mongoose.Types.ObjectId.isValid(id)) {
-    res.status(400).json({ error: "No such Folder" });
-    return;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      res.status(400).json({ error: "No such Folder" });
+      return;
+    }
+
+    const folder = await Folder.findOneAndDelete({ _id: id, user });
+
+    if (!folder) {
+      res.status(404).json({ error: "No such Folder" });
+      return;
+    }
+
+    await Note.deleteMany({ folder: folder.id });
+    res.status(202).json(folder);
   }
-
-  const folder = await Folder.findByIdAndDelete(id);
-
-  if (!folder) {
-    res.status(404).json({ error: "No such Folder" });
-    return;
-  }
-
-  await Note.deleteMany({ folder: folder.id });
-  res.status(202).json(folder);
 };
 
 // Update a Folder
@@ -84,20 +97,23 @@ export const patchFolder: RequestHandler = async (
   req: Request,
   res: Response
 ) => {
-  const { id } = req.params;
-  const { notes } = req.body as FolderObj;
+  const { user } = req;
+  if (user) {
+    const { id } = req.params;
+    const { notes } = req.body as FolderObj;
 
-  if (!mongoose.Types.ObjectId.isValid(id)) {
-    res.status(404).json({ error: "No such Folder" });
-    return;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      res.status(404).json({ error: "No such Folder" });
+      return;
+    }
+
+    const folder = await Folder.findOneAndUpdate({ _id: id, user }, { notes });
+
+    if (!folder) {
+      res.status(404).json({ error: "No such Folder" });
+      return;
+    }
+
+    res.status(200).json(folder);
   }
-
-  const folder = await Folder.findByIdAndUpdate(id, { notes });
-
-  if (!folder) {
-    res.status(404).json({ error: "No such Folder" });
-    return;
-  }
-
-  res.status(200).json(folder);
 };
